// packages/uberstrike-js/src/UberStrike/Core/Models/CommActorInfo.ts
/*
 * Copyright (C) 2017, 2021-2024 Team FESTIVAL
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

import { ChannelType, MemberAccessLevel } from '@/Cmune/DataCenter/Common/Entities';
import GameRoom from './GameRoom';

export default class CommActorInfo {
  Cmid: number;
  PlayerName: string;
  AccessLevel: MemberAccessLevel;
  Channel: ChannelType;
  ClanTag: string;
  CurrentRoom: GameRoom;
  ModerationFlag: number;
  ModInformation: string;

  constructor(params: Partial<CommActorInfo> = {}) {
    Object.assign(this, params);
  }
}